requirements.txt
deepgram-sdk
python-dotenv
Flask
flask-cors
groq 
